reworked Mail system
