+++ src/server/game/Calendar/CalendarMgr.cpp
@@ -22,8 +22,7 @@
 #include "Guild.h"
 #include "GuildMgr.h"
 #include "Log.h"
-#include "Mail.h"
-#include "Mail.h"
+#include "MailMgr.h"
 #include "ObjectAccessor.h"
 #include "Opcodes.h"
 #include "Player.h"
