+++ src/server/scripts/Northrend/zone_dalaran.cpp
@@ -25,8 +25,7 @@ Script Data End */
 
 #include "ScriptMgr.h"
 #include "DatabaseEnv.h"
-#include "Mail.h"
-#include "Mail.h"
+#include "MailMgr.h"
 #include "Map.h"
 #include "MotionMaster.h"
 #include "Player.h"
